airflow-pipeline/dags/requirements.txt
alembic==1.4.2
amqp==2.5.2
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
cattrs==0.9.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
cryptography==2.9
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
flower==0.9.4
fsspec==0.7.2
funcsigs==1.0.2
future==0.16.0
graphviz==0.13.2
gunicorn==19.10.0
hmsclient==0.1.1
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
iso8601==0.1.12
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10.3
jmespath==0.9.5
JPype1==0.7.2
json-merge-patch==0.2
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.14
ndg-httpsclient==0.5.1
numpy==1.18.2
pandas==0.25.3
paramiko==2.7.1
pendulum==1.4.4
prison==0.1.3
psutil==5.7.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyHive==0.6.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pysftp==0.2.9
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-snappy==0.5.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
redis==3.2.0
requests==2.23.0
setproctitle==1.1.10
six==1.14.0
SQLAlchemy==1.3.15
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.3
sshtunnel==0.1.5
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
tornado==5.1.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==0.16.1
WTForms==2.2.1
zipp==3.1.0
zope.deprecation==4.4.0
oauth2client
pandasql
ndjson